Reimplement the j8rt test using a tiny stub application which forces analysis into the problematic JDK code.
We don't intent to actually test anything about the Stub itself, it is just
used to include JDK code into the callgraph which are the interesting pieces
we test. This allow the j8rt tests to work just like any normal application
as far as the test infrastructure is concerned.