Remove all the logging code
The log file was mostly redundant with pytests own reporting, and I
actually found it somewhat annoying that stderr would only go to
the log file (enough that I just aways cat'ed the log file after running).

Instead now we let pytest capture stderr normally (which means it will
be reported in pytest's own report when a command fails).

Initially *I* actually advocatd for the log file, due to the noisiness of
errors. However, since then we've cleaned up the testing infrastructure and
no longer have cascading failures (that is, if an adapter fails we don't
also try and fail to run all the tests with it).

Also cleanup utils.run_cmd a bit. It was always used in a context where
the return code was checked, so just check it in run_cmd itself, so all
clients don't have to. Additionally, as we're letting pytest capture stderr,
don't read it ourselves and return it to the caller.